==> Nanostring/Nanostring_analysis/R Code/07_Gene_enrichment.R <==
# 07 Gene Enrichment 
# Paul Parodi



######### GOAL ##############
# Show gene enrichment of our data

##### LIBRARY ###########
#### LOADING IN THE FILES ###########
##### Gene enrichment function ########
##### Calling gene enrichment #########
##### Visualizations ##########
###### EXPORT #########
